Improve the user documentation.
This commit fixes a couple of typos in the Javadoc, which resulted
from an automatic refactor process done in IntelliJ.  Documentation
should be more straightforward for users now!

Secondarily, the base README now includes notes about how to perform
releases and documentation updates if the engineer is somewhat
unfamiliar with Maven.