Enhance clarity in notes about metric export.
The original Metric export documentation was a little ambiguous in
parts.  This commit strengthens the optionalness of the use of the
runtime annotations but notes the legitimate use cases of it whereby
the Virtual Machine Specification will fall short and neither load nor
initialize classes of interest that may have metrics defined that an
application's consumer may care about.

This is likely to be a corner case that most users of Prometheus will
hopefully never encounter.  One cannot be too sure, though, with how
some of dynamic languages built on top of the Java Virtual Machine
generate their output classes, so this mechanism is offered as a
fallback.

In short, using this annotation is a good and safe practice.  Not
using it is fine in most cases but can put you at the perils of the
VMS and its loader behaviors.