Reference the Google Java Style guide.
This style guide has been pretty well etched into my mind for the last
seven years, and it is what I have used for all of my own projects,
including the Java components of Prometheus.  Given that we have new
contributors, we should offer them the courtesy of letting them know
in advance what the styles and patterns are.