Skip exposition of Summary children w/o samples.
Issue https://github.com/prometheus/client_java/issues/27 reported
a case whereby the quantile estimation library would throw an
IllegalStateException for the request of quantile ranks for observation
streams that are empty.  This was a fix for bad upstream behavior of
original estimator library that would either throw a
NullPointerException or return a zero sentinel value depending upon
what condition happened under the hood.  Nevertheless, the original
fix was insufficient for the client library.

What has been done is provide a filter mechanism in exposition to
prevent serialization of empty Summary metrics.  Along with this
has been an upgrade of the quantile estimator library so it no
longer throws an exception.