Fix notes about thread safety and mutability.
Erroneously some of the Javadoc for Metric.Partial and its derivatives
said that references to said types should be avoided at all costs.
This statement should have said that retaining references to these
are fine, so long as one realizes they are mutable types and not
thread safe.

Metric.Child and its derivatives have had their documentation enhanced
such that it now says that references onto these should not be help but
rather references onto Metric.Partial and friends are OK.

To that note, the synchronization and waiting parameters in the
JvmstatMonitor class have been fixed, for they were wrong originally.