Add describe and autodescribe to detect dupes. (#162)
This works largely the same as this feature in the
Python client.

This works by adding an optional Describable interface to collectors,
which returns data in the same format as collect (though hopefully
without the samples). If implemented it is called at registration time.

If describe is not present and auto describe is set on the
registry, then collect is called instead. This is enabled by default on
the default registry, but disabled elsewhere.

Put in empty describes on some custom collectors that
deal with arbitrary data.