SimpleCollecor.labels() initialization performance (#307)
- Now, we do 3 lookups in the children Map when we create a new tupple of
labels. From my understanding, it's to manage concurrent initializations.

It can be improved. I suggest to check the result of 'putIfAbsent' to avoid the third
lookup.

Because we do not want to create a new Child for each call to #labels, I
don't find a solution with only 1 lookup. BUT with Java 8, we can consider
the use of 'computeIfAbsent' with a lambda.