More helpful error msg calling labelled collector with no labels
Calling a labelled collector without labels threw a NullPointerException without any explanation of the reason for the error. Change this to throw IllegalArgumentException with an explanatory message. This is consistent with the error thrown when calling a labelled collector with a mismatching number of labels.

Fixes https://github.com/prometheus/client_java/issues/332