Fix overlap between invoke() and value() in Spring MVC
I found that sometimes when I hit the /prometheus endpoint on my Spring
Boot application, I got a blank page back. Other times it returned the
expected set of complete results.

After investigation, I think the problem is that
EndpointMvcAdapter.invoke() has a default @RequestMapping that matches
the @RequestMapping of the value(name) method defined on
PrometheusMvcEndpoint.

Defining a distinct "/filtered" request mapping on the value method
allows filtered results to consistently work by hitting the URL
/prometheus/filtered?name[]=foo , while not having the issue of
unfiltered results sometimes not working when hitting /prometheus.

It was necessary to override "invoke" and add an explicit mapping there
as well, or the unit tests that already existed showed that a JSON
content type was being returned by Spring due to the definition of
@ActuatorGetMapping in Spring Boot 1.5.x.

For whatever reason, with my version of Spring Boot (1.3.7) and Spring
Framework (4.2.7), the defaultValue of "" for value does not turn into
an empty Set. This does work with the version of Spring Boot that the
tests use, but change thus also improves compatibility with older spring
versions.

Signed-off-by: Patrick Cavanaugh <pcavanaugh@vistaprint.com>