Adjustment to allow invoke() to work with any Accept header
Testing revealed that a missing accept header would result in Spring
returning a 406 when invoke() was called; this behavior did not match
the previous behavior of value().

Changed the structure of invoke() to make it behave as close as possible
to value() by using the same @ResponseBody with the same response header
content type, and the same "produces" attribute in the request mapping.

Signed-off-by: Patrick Cavanaugh <pcavanaugh@vistaprint.com>