Make Collector labels-to-child map implementation pluggable
#460 proposed an optimized zero-GC version of the child lookup logic
which was deemed too specialized for inclusion in the library.

Subjectively less complex alternatives were also proposed which provide
some but not as much performance/garbage improvement, and rely for
example on some per-thread overhead.

This PR aims to add a minimally-invasive mechanism to allow users to
plug in an implementation of their choice, so that performance sensitive
consumers can opt for minimal overhead without the core library having
to include the corresponding code.

Signed-off-by: nickhill <nickhill@us.ibm.com>