Add automatic module names for compatibility with Jigsaw
This adds entries to the JAR manifest files to allow consumption of Prometheus
by the Java module system (JPMS aka Jigsaw) introduced in Java 9. Since these
are added only to the manifests, rather than the `module-info.class`, they will
not change behaviour for JVMs that do not use the module system. Some modules
are not compliant with the requirement that a package be present in only one
module, so they are excluded.

Signed-off-by: Andre Masella <andre@masella.name>