Let DropwizardExports users deal with snapshot validation errors
In the current form, if e.g. a Counter contains invalid data the entire `collect` calls fails with
InvalidArgumentException.

In large applications, it may be preferable to gracefully degrade the available metrics and report
the failure via e.g. logging. This change allows that by letting the user add a exception handler.

This allows the user to inspect the exception, and the metric name it was associated with, and
decide if the exception should be suppressed or not.

Metrics for which exceptions were suppressed will be absent from the collection.

Signed-off-by: Anders Månsson <anders.mansson@neo4j.com>