Allow subclasses of DropwizardExports to deal with validation errors
In the current form, if e.g. a Counter contains invalid data the entire
`collect` calls fails with InvalidArgumentException.

In large applications, it may be preferable to gracefully degrade the
available metrics and report the failure via e.g. logging. This change
allows just that - by changing the access modifiers to the `fromXYZ` methods
subclasses can now handle the errors.

Finally, to allow for skipping of invalid metrics, we follow the example
in `fromGauge` to interpret the meaning of a returned null to be that the
metric should be skipped.