Add metricsHandlerPath and registerHealthHandler configuration options to HTTPServer (#1735)
Hey @fstab, long time no see.

We would like to use this exporter in our project, but need to customize
the metrics endpoint and disable the health handler. I hope this small
contribution can make into a release.

## Summary

This pull request enhances the configurability of the Prometheus Java
HTTP server by allowing customization of the metrics endpoint path and
making the health endpoint registration optional. It also updates the
default HTML handler to reflect these changes and adds tests to ensure
correct behavior.

### Configuration improvements

* Added support for customizing the metrics endpoint via the new
`metricsHandlerEndpoint` option in the `HTTPServer.Builder`. The default
remains `/metrics`, but it can now be changed as needed.
[[1]](diffhunk://#diff-fd57f399229770ac99d7f2b18cee74741a3f70c7c4c0800e17a04cb1b1bc321eL60-R82)
[[2]](diffhunk://#diff-fd57f399229770ac99d7f2b18cee74741a3f70c7c4c0800e17a04cb1b1bc321eR187-R191)
[[3]](diffhunk://#diff-fd57f399229770ac99d7f2b18cee74741a3f70c7c4c0800e17a04cb1b1bc321eR264-R278)
[[4]](diffhunk://#diff-fd57f399229770ac99d7f2b18cee74741a3f70c7c4c0800e17a04cb1b1bc321eL278-R302)
* Introduced a `registerHealthHandler` option in the
`HTTPServer.Builder` to control whether the health endpoint
(`/-/healthy`) is registered. By default, it is enabled.
[[1]](diffhunk://#diff-fd57f399229770ac99d7f2b18cee74741a3f70c7c4c0800e17a04cb1b1bc321eL60-R82)
[[2]](diffhunk://#diff-fd57f399229770ac99d7f2b18cee74741a3f70c7c4c0800e17a04cb1b1bc321eR187-R191)
[[3]](diffhunk://#diff-fd57f399229770ac99d7f2b18cee74741a3f70c7c4c0800e17a04cb1b1bc321eR264-R278)
[[4]](diffhunk://#diff-fd57f399229770ac99d7f2b18cee74741a3f70c7c4c0800e17a04cb1b1bc321eL278-R302)

### Default handler updates

* Updated the `DefaultHandler` to accept the metrics path as a parameter
and generate all documentation links and examples based on the
configured path, ensuring consistency in the UI and documentation.
[[1]](diffhunk://#diff-99311d9bc5b7879fc572852eb8106a440466befcdc072a07632818e6425dbb07L14-R34)
[[2]](diffhunk://#diff-99311d9bc5b7879fc572852eb8106a440466befcdc072a07632818e6425dbb07L53-R66)

### Testing

* Added tests to verify that the metrics endpoint can be customized and
that the health endpoint can be enabled or disabled as configured.
[[1]](diffhunk://#diff-c5c2cdc04572f6033ea66e1b7cd4859178b7da6adea0f8fc46993ae83ce8fa6cR108-R120)
[[2]](diffhunk://#diff-c5c2cdc04572f6033ea66e1b7cd4859178b7da6adea0f8fc46993ae83ce8fa6cR163-R188)

Signed-off-by: David Sondermann <david.sondermann@hivemq.com>